don't assume sdk install anymore, reduce dead code /  duplicate impls
